Deduplicate scaled_mm_allowed_device into ScaledBlasUtils
_scaled_mm_allowed_device was copy-pasted in cuda/ScaledBlas.cpp and
cuda/GroupedBlas.cpp with identical logic. That creates a maintenance
risk: a future device-capability change could be applied in one file
and missed in the other.
Move the helper to at::native::scaled as scaled_mm_allowed_device,
declared in ScaledBlasUtils.h and defined once in
cuda/ScaledBlasDeviceUtils.cpp. Update both call sites to use the
shared helper.

This intentionally extends gfx1250 support to torch._scaled_grouped_mm's
device gate (_scaled_grouped_mm_cuda, _scaled_grouped_mm_cuda_v2, and the
bf16 fast-path check in GroupedBlas.cpp), aligning it with the gate
already used by torch._scaled_mm. Confirmed as the intended behavior.
